Split module up
