+++ bin/validator/src/commands/bootstrap.rs
@@ -1,0 +1,86 @@
+use std::path::{Path, PathBuf};
+
+use anyhow::Context;
+use miden_node_store::genesis::config::{AccountFileWithName, GenesisConfig};
+use miden_node_utils::fs::ensure_empty_directory;
+use miden_node_utils::signer::BlockSigner;
+use miden_node_validator::ValidatorSigner;
+use miden_protocol::utils::serde::Serializable;
+
+use super::ValidatorKey;
+
+const GENESIS_BLOCK_FILENAME: &str = "genesis.dat";
+
+pub async fn run(
+    genesis_block_directory: &Path,
+    accounts_directory: &Path,
+    data_directory: &Path,
+    genesis_config: Option<&PathBuf>,
+    validator_key: ValidatorKey,
+) -> anyhow::Result<()> {
+    let config = genesis_config
+        .map(|file_path| {
+            GenesisConfig::read_toml_file(file_path).with_context(|| {
+                format!("failed to parse genesis config from file {}", file_path.display())
+            })
+        })
+        .transpose()?
+        .unwrap_or_default();
+
+    for directory in [accounts_directory, genesis_block_directory] {
+        ensure_empty_directory(directory)?;
+    }
+
+    let signer = validator_key.into_signer().await?;
+    match signer {
+        ValidatorSigner::Kms(signer) => {
+            build_and_write_genesis(config, signer, accounts_directory, genesis_block_directory, data_directory).await
+        },
+        ValidatorSigner::Local(signer) => {
+            build_and_write_genesis(config, signer, accounts_directory, genesis_block_directory, data_directory).await
+        },
+    }
+}
+
+/// Builds the genesis state, writes account secret files, signs the genesis block, writes it
+/// to disk, and initializes the validator's database with the genesis block as the chain tip.
+async fn build_and_write_genesis(
+    config: GenesisConfig,
+    signer: impl BlockSigner,
+    accounts_directory: &Path,
+    genesis_block_directory: &Path,
+    data_directory: &Path,
+) -> anyhow::Result<()> {
+    let (genesis_state, secrets) = config.into_state(signer)?;
+
+    for item in secrets.as_account_files(&genesis_state) {
+        let AccountFileWithName { account_file, name } = item?;
+        let account_path = accounts_directory.join(name);
+        // Do not override existing keys.
+        fs_err::OpenOptions::new()
+            .create_new(true)
+            .write(true)
+            .open(&account_path)
+            .context("key file already exists")?;
+        account_file.write(account_path)?;
+    }
+
+    let genesis_block =
+        genesis_state.into_block().await.context("failed to build the genesis block")?;
+
+    let block_bytes = genesis_block.inner().to_bytes();
+    let genesis_block_path = genesis_block_directory.join(GENESIS_BLOCK_FILENAME);
+    fs_err::write(&genesis_block_path, block_bytes).context("failed to write genesis block")?;
+
+    let (genesis_header, ..) = genesis_block.into_inner().into_parts();
+    let db = miden_node_validator::db::load(data_directory.join("validator.sqlite3"))
+        .await
+        .context("failed to initialize validator database during bootstrap")?;
+    db.transact("upsert_block_header", move |conn| {
+        miden_node_validator::db::upsert_block_header(conn, &genesis_header)
+    })
+    .await
+    .context("failed to persist genesis block header as chain tip")?;
+
+    Ok(())
+}
