+++ bin/validator/src/main.rs
@@ -1,30 +1,7 @@
-use std::net::SocketAddr;
-use std::path::{Path, PathBuf};
-
-use anyhow::Context;
-use clap::{Parser, Subcommand};
-use miden_node_store::genesis::config::{AccountFileWithName, GenesisConfig};
-use miden_node_utils::clap::GrpcOptionsInternal;
-use miden_node_utils::fs::ensure_empty_directory;
-use miden_node_utils::grpc::UrlExt;
+use clap::Parser;
 use miden_node_utils::logging::OpenTelemetry;
-use miden_node_utils::signer::BlockSigner;
-use miden_node_validator::{Validator, ValidatorSigner};
-use miden_protocol::crypto::dsa::ecdsa_k256_keccak::SecretKey;
-use miden_protocol::utils::serde::{Deserializable, Serializable};
-use url::Url;
 
-const ENV_DATA_DIRECTORY: &str = "MIDEN_NODE_DATA_DIRECTORY";
-const ENV_ENABLE_OTEL: &str = "MIDEN_NODE_ENABLE_OTEL";
-const ENV_URL: &str = "MIDEN_NODE_VALIDATOR_URL";
-const ENV_GENESIS_CONFIG_FILE: &str = "MIDEN_NODE_VALIDATOR_GENESIS_CONFIG_FILE";
-const ENV_KEY: &str = "MIDEN_NODE_VALIDATOR_KEY";
-const ENV_KMS_KEY_ID: &str = "MIDEN_NODE_VALIDATOR_KMS_KEY_ID";
-
-/// A predefined, insecure validator key for development purposes.
-const INSECURE_KEY_HEX: &str = "0101010101010101010101010101010101010101010101010101010101010101";
-/// The filename used for the genesis block file.
-const GENESIS_BLOCK_FILENAME: &str = "genesis.dat";
+mod commands;
 
 // CLI
 // ================================================================================================
@@ -33,7 +10,7 @@ const GENESIS_BLOCK_FILENAME: &str = "genesis.dat";
 #[command(version, about, long_about = None)]
 struct Cli {
     #[command(subcommand)]
-    command: ValidatorCommand,
+    command: commands::ValidatorCommand,
 }
 
 impl Cli {
@@ -43,276 +20,6 @@ impl Cli {
         } else {
             OpenTelemetry::Disabled
         }
-    }
-}
-
-// COMMANDS
-// ================================================================================================
-
-#[derive(Subcommand)]
-enum ValidatorCommand {
-    /// Bootstraps the genesis block.
-    ///
-    /// Creates accounts from the genesis configuration, builds and signs the genesis block,
-    /// and writes the signed block and account secret files to disk. Also initializes the
-    /// validator's database with the genesis block as the chain tip.
-    Bootstrap {
-        /// Directory in which to write the genesis block file.
-        #[arg(long, value_name = "DIR")]
-        genesis_block_directory: PathBuf,
-        /// Directory to write the account secret files (.mac) to.
-        #[arg(long, value_name = "DIR")]
-        accounts_directory: PathBuf,
-        /// Directory in which to store the validator's database.
-        #[arg(long, env = ENV_DATA_DIRECTORY, value_name = "DIR")]
-        data_directory: PathBuf,
-        /// Use the given configuration file to construct the genesis state from.
-        #[arg(long, env = ENV_GENESIS_CONFIG_FILE, value_name = "GENESIS_CONFIG")]
-        genesis_config_file: Option<PathBuf>,
-        /// Configuration for the Validator key used to sign the genesis block.
-        #[command(flatten)]
-        validator_key: ValidatorKey,
-    },
-
-    /// Starts the validator component.
-    Start {
-        /// Url at which to serve the gRPC API.
-        #[arg(env = ENV_URL)]
-        url: Url,
-
-        /// Enables the exporting of traces for OpenTelemetry.
-        ///
-        /// This can be further configured using environment variables as defined in the official
-        /// OpenTelemetry documentation. See our operator manual for further details.
-        #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
-        enable_otel: bool,
-
-        #[command(flatten)]
-        grpc_options: GrpcOptionsInternal,
-
-        /// Directory in which to store the validator's data.
-        #[arg(long, env = ENV_DATA_DIRECTORY, value_name = "DIR")]
-        data_directory: PathBuf,
-
-        /// Insecure, hex-encoded validator secret key for development and testing purposes.
-        ///
-        /// If not provided, a predefined key is used.
-        ///
-        /// Cannot be used with `key.kms-id`.
-        #[arg(
-            long = "key.hex",
-            env = ENV_KEY,
-            value_name = "VALIDATOR_KEY",
-            default_value = INSECURE_KEY_HEX,
-            group = "key"
-        )]
-        validator_key: String,
-
-        /// Key ID for the KMS key used by validator to sign blocks.
-        ///
-        /// Cannot be used with `key.hex`.
-        #[arg(
-            long = "key.kms-id",
-            env = ENV_KMS_KEY_ID,
-            value_name = "VALIDATOR_KMS_KEY_ID",
-            group = "key"
-        )]
-        kms_key_id: Option<String>,
-    },
-}
-
-impl ValidatorCommand {
-    async fn handle(self) -> anyhow::Result<()> {
-        match self {
-            Self::Bootstrap {
-                genesis_block_directory,
-                accounts_directory,
-                data_directory,
-                genesis_config_file,
-                validator_key,
-            } => {
-                bootstrap_genesis(
-                    &genesis_block_directory,
-                    &accounts_directory,
-                    &data_directory,
-                    genesis_config_file.as_ref(),
-                    validator_key,
-                )
-                .await
-            },
-            Self::Start {
-                url,
-                grpc_options,
-                validator_key,
-                data_directory,
-                kms_key_id,
-                ..
-            } => {
-                let address = url
-                    .to_socket()
-                    .context("failed to extract socket address from validator URL")?;
-
-                if let Some(kms_key_id) = kms_key_id {
-                    let signer = ValidatorSigner::new_kms(kms_key_id).await?;
-                    serve(address, grpc_options, signer, data_directory).await
-                } else {
-                    let signer = SecretKey::read_from_bytes(hex::decode(validator_key)?.as_ref())?;
-                    let signer = ValidatorSigner::new_local(signer);
-                    serve(address, grpc_options, signer, data_directory).await
-                }
-            },
-        }
-    }
-
-    fn is_open_telemetry_enabled(&self) -> bool {
-        match self {
-            Self::Start { enable_otel, .. } => *enable_otel,
-            Self::Bootstrap { .. } => false,
-        }
-    }
-}
-
-// VALIDATOR KEY
-// ================================================================================================
-
-/// Configuration for the Validator key used to sign blocks.
-#[derive(clap::Args)]
-#[group(required = false, multiple = false)]
-struct ValidatorKey {
-    /// Insecure, hex-encoded validator secret key for development and testing purposes.
-    ///
-    /// If not provided, a predefined key is used.
-    ///
-    /// Cannot be used with `validator.key.kms-id`.
-    #[arg(
-        long = "validator.key.hex",
-        env = ENV_KEY,
-        value_name = "VALIDATOR_KEY",
-        default_value = INSECURE_KEY_HEX,
-    )]
-    validator_key: String,
-    /// Key ID for the KMS key used by validator to sign blocks.
-    ///
-    /// Cannot be used with `validator.key.hex`.
-    #[arg(
-        long = "validator.key.kms-id",
-        env = ENV_KMS_KEY_ID,
-        value_name = "VALIDATOR_KMS_KEY_ID",
-    )]
-    validator_kms_key_id: Option<String>,
-}
-
-impl ValidatorKey {
-    async fn into_signer(self) -> anyhow::Result<ValidatorSigner> {
-        if let Some(kms_key_id) = self.validator_kms_key_id {
-            Ok(ValidatorSigner::new_kms(kms_key_id).await?)
-        } else {
-            let signer = SecretKey::read_from_bytes(hex::decode(self.validator_key)?.as_ref())?;
-            Ok(ValidatorSigner::new_local(signer))
-        }
-    }
-}
-
-// HELPERS
-// ================================================================================================
-
-async fn serve(
-    address: SocketAddr,
-    grpc_options: GrpcOptionsInternal,
-    signer: ValidatorSigner,
-    data_directory: PathBuf,
-) -> anyhow::Result<()> {
-    Validator { address, grpc_options, signer, data_directory }
-        .serve()
-        .await
-        .context("failed while serving validator component")
-}
-
-/// Builds the genesis state, writes account secret files, signs the genesis block, writes it
-/// to disk, and initializes the validator's database with the genesis block as the chain tip.
-async fn build_and_write_genesis(
-    config: GenesisConfig,
-    signer: impl BlockSigner,
-    accounts_directory: &Path,
-    genesis_block_directory: &Path,
-    data_directory: &Path,
-) -> anyhow::Result<()> {
-    let (genesis_state, secrets) = config.into_state(signer)?;
-
-    for item in secrets.as_account_files(&genesis_state) {
-        let AccountFileWithName { account_file, name } = item?;
-        let account_path = accounts_directory.join(name);
-        // Do not override existing keys.
-        fs_err::OpenOptions::new()
-            .create_new(true)
-            .write(true)
-            .open(&account_path)
-            .context("key file already exists")?;
-        account_file.write(account_path)?;
-    }
-
-    let genesis_block =
-        genesis_state.into_block().await.context("failed to build the genesis block")?;
-
-    let block_bytes = genesis_block.inner().to_bytes();
-    let genesis_block_path = genesis_block_directory.join(GENESIS_BLOCK_FILENAME);
-    fs_err::write(&genesis_block_path, block_bytes).context("failed to write genesis block")?;
-
-    let (genesis_header, ..) = genesis_block.into_inner().into_parts();
-    let db = miden_node_validator::db::load(data_directory.join("validator.sqlite3"))
-        .await
-        .context("failed to initialize validator database during bootstrap")?;
-    db.transact("upsert_block_header", move |conn| {
-        miden_node_validator::db::upsert_block_header(conn, &genesis_header)
-    })
-    .await
-    .context("failed to persist genesis block header as chain tip")?;
-
-    Ok(())
-}
-
-async fn bootstrap_genesis(
-    genesis_block_directory: &Path,
-    accounts_directory: &Path,
-    data_directory: &Path,
-    genesis_config: Option<&PathBuf>,
-    validator_key: ValidatorKey,
-) -> anyhow::Result<()> {
-    let config = genesis_config
-        .map(|file_path| {
-            GenesisConfig::read_toml_file(file_path).with_context(|| {
-                format!("failed to parse genesis config from file {}", file_path.display())
-            })
-        })
-        .transpose()?
-        .unwrap_or_default();
-
-    for directory in [accounts_directory, genesis_block_directory] {
-        ensure_empty_directory(directory)?;
-    }
-
-    let signer = validator_key.into_signer().await?;
-    match signer {
-        ValidatorSigner::Kms(signer) => {
-            build_and_write_genesis(
-                config,
-                signer,
-                accounts_directory,
-                genesis_block_directory,
-                data_directory,
-            )
-            .await
-        },
-        ValidatorSigner::Local(signer) => {
-            build_and_write_genesis(
-                config,
-                signer,
-                accounts_directory,
-                genesis_block_directory,
-                data_directory,
-            )
-            .await
-        },
     }
 }
